Rework Style
Particles demo was causing issues for people, reworked this to be more compatible.

There's still an issue of this not aligning properly when scaling and likely on mobile devices. We need another pass on  this.

I'd be fine pushing out this current version since it alleviates the problem of the particles demo for now.